New ui package.
Contains only the original CLI for now.
