+++ opalescence/ui/__init__.py
@@ -1,0 +1,6 @@
+#!/usr/bin/env python
+# -*- coding: utf-8 -*-
+"""
+The UI for opalescence is implemented here.
+Currently, only a cli is provided.
+"""
